refactor: separate components into folders
