+++ src/stories/header-template/Header.tsx
@@ -1,6 +1,6 @@
 import React from "react"
 
-import { Button } from "./Button"
+import { Button } from "../button-template/Button"
 import "./header.css"
 
 type User = {
