+++ src/stories/page-template/Page.tsx
@@ -1,6 +1,6 @@
 import React from "react"
 
-import { Header } from "./Header"
+import { Header } from "../header-template/Header"
 import "./page.css"
 
 type User = {
